examples/python/openrouter-tool-agent/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

traceroot==0.1.4
